refactor(database): Change Habit ID type to Long and update HabitDao
Updates the `Habit` entity and `HabitDao` to use `Long` for the primary key (`habitId`) instead of `Int`. This provides a wider range for auto-generated IDs.

- Updates the `habitId` type in the `Habit` data class to `Long`.
- Modifies `HabitDao` methods (`insertHabit`, `getHabitById`, `getHabitByIdFlow`, `setHabitInactive`) to accept or return `Long` for the habit ID.
- Adjusts instrumented tests in `HabitDaoTest` to reflect the change in ID type and handle inserting and retrieving the `Long` ID correctly.

+++ app/src/main/java/io/track/habit/domain/model/database/Habit.kt
@@ -17,7 +17,7 @@ import java.util.Date
  */
 @Entity(tableName = "habits")
 data class Habit(
-    @PrimaryKey(autoGenerate = true) val habitId: Int = 0,
+    @PrimaryKey(autoGenerate = true) val habitId: Long = 0,
     val name: String,
     val isActive: Boolean = true,
     val createdAt: Date = Date(),
